+++ cli/cli/commands/exec/exec.go
@@ -16,6 +16,7 @@ import (
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"os"
+	"path/filepath"
 	"strconv"
 	"strings"
 )
@@ -242,7 +243,15 @@ func executeScript(enclaveCtx *enclaves.EnclaveContext, scriptPath string, dryRu
 }
 
 func executeModule(enclaveCtx *enclaves.EnclaveContext, modulePath string, serializedParams string, dryRun bool) error {
-	executionResponse, err := enclaveCtx.ExecuteStartosisModule(modulePath, serializedParams, dryRun)
+	// we get the absolute path so that the logs make more sense
+	absoluteModulePath, err := filepath.Abs(modulePath)
+	logrus.Infof("Executing Starlark package at '%v' as the passed argument '%v' looks like a directory", absoluteModulePath, modulePath)
+
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred while getting the absolute path for '%v'", modulePath)
+	}
+
+	executionResponse, err := enclaveCtx.ExecuteStartosisModule(absoluteModulePath, serializedParams, dryRun)
 	if err != nil {
 		return stacktrace.Propagate(err, "An unexpected error occurred executing the Startosis module '%s'", modulePath)
 	}
@@ -318,7 +327,7 @@ func getOrCreateEnclaveContext(
 		}
 		return enclaveContext, false, nil
 	}
-	logrus.Infof("Creating a new enclave for the startosis script to execute inside...")
+	logrus.Infof("Creating a new enclave for Starlark to execute inside...")
 	enclaveContext, err := kurtosisContext.CreateEnclave(ctx, enclaveId, isPartitioningEnabled)
 	if err != nil {
 		return nil, false, stacktrace.Propagate(err, fmt.Sprintf("Unable to create new enclave with ID '%s'", enclaveId))